change the tool for Json, add behavior for moving the test up and down, add button to quit

+++ Assets/Scripts/EditWindowController.cs
@@ -12,7 +12,7 @@ public class EditWindowController : MonoBehaviour
     private Color defaultColor;
     private List<Image> testButtonImages = new List<Image>();
 
-    public static string NameOfSelectedTest;
+    public static string NameOfSelectedTest = "";
 
     private void Start()
     {
@@ -22,14 +22,14 @@ public class EditWindowController : MonoBehaviour
     private void Init()
     {
         Clear();
-        var tests = DataParser.GetNamesOfTests();
+        var tests = DataParser.GetAllTests();
         if (tests.Count == 0)
             return;
 
-        foreach (var word in tests)
+        for (int i = 0; i < tests.Count - 1; i++)
         {
             var wordGO = Instantiate(textObj, place);
-            wordGO.transform.GetComponentInChildren<Text>().text = word;
+            wordGO.transform.GetComponentInChildren<Text>().text = tests[i];
             wordGO.GetComponent<Button>().onClick.AddListener(OnTestButtonClick);
             testButtonImages.Add(wordGO.GetComponent<Image>());
         }
@@ -43,6 +43,8 @@ public class EditWindowController : MonoBehaviour
 
         for (int i = 0; i < place.childCount; i++)
             Destroy(place.GetChild(i).gameObject);
+
+        NameOfSelectedTest = "";
     }
 
     public void OnTestButtonClick()
@@ -68,7 +70,90 @@ public class EditWindowController : MonoBehaviour
     public void OnDeleteTestButtonClicked()
     {
         var clickedButton = EventSystem.current.currentSelectedGameObject;
+        if (NameOfSelectedTest.Length == 0)
+            return;
+
+        DataParser.DeleteTest(NameOfSelectedTest);
         DataParser.DeleteFile(NameOfSelectedTest);
         Init();
+    }
+
+    public void OnMoveUpButtonClicked()
+    {
+        var clickedButton = EventSystem.current.currentSelectedGameObject;
+        if (clickedButton.name.Equals("Move Up"))
+        {
+            var tests = DataParser.GetAllTests();
+            if (tests.Count == 0)
+                return;
+
+            string selectedTest = "";
+            for (int i = 0; i < place.childCount; i++)
+                if (place.GetChild(i).GetComponent<Image>().color == selectedColor)
+                {
+                    selectedTest = place.GetChild(i).GetComponentInChildren<Text>().text;
+                    break;
+                }
+
+            if (selectedTest.Length == 0)
+                return;
+
+            var testsInNewOrder = new List<string>();
+            for (int i = 0; i < tests.Count - 1; i++)
+            {
+                if (tests[i + 1].Equals(selectedTest))
+                {
+                    testsInNewOrder.Add(tests[i + 1]);
+                    testsInNewOrder.Add(tests[i]);
+                    i++;
+                    continue;
+                }
+                testsInNewOrder.Add(tests[i]);
+            }
+
+            DataParser.WriteAllTests(testsInNewOrder);
+
+            Init();
+        }
+    }
+
+    public void OnMoveDownButtonClicked()
+    {
+        var clickedButton = EventSystem.current.currentSelectedGameObject;
+        if (clickedButton.name.Equals("Move Down"))
+        {
+            var tests = DataParser.GetAllTests();
+            if (tests.Count == 0)
+                return;
+
+            string selectedTest = "";
+            for (int i = 0; i < place.childCount; i++)
+                if (place.GetChild(i).GetComponent<Image>().color == selectedColor)
+                {
+                    selectedTest = place.GetChild(i).GetComponentInChildren<Text>().text;
+                    break;
+                }
+
+            if (selectedTest.Length == 0 || 
+                tests.IndexOf(selectedTest) == tests.Count - 2)
+                return;
+
+            var testsInNewOrder = new List<string>();
+            for (int i = 0; i < tests.Count - 1; i++)
+            {
+                if (tests[i].Equals(selectedTest))
+                {
+                    testsInNewOrder.Add(tests[i + 1]);
+                    testsInNewOrder.Add(tests[i]);
+                    i++;
+                    continue;
+                }
+                testsInNewOrder.Add(tests[i]);
+            }
+
+            DataParser.WriteAllTests(testsInNewOrder);
+
+            Init();
+        }
     }
 }
